Use page objects in IcrcTransactionCard and -List tests (#3700)
# Motivation

Tests with page objects are easier to read and maintain.
I will be changing these tests when making UI changes for ckBTC
transactions.

# Changes

1. Extend `createIcrcTransactionWithId` to make `to` and `from` params
optional and accept additional optional `amount` and `fee` params.
2. Add `getHeadline` and `getDate` and static `under` to
`TransactionCardPo` page object.
3. Use page objects in `IcrcTransactionCard` and `IcrcTransactionsList`.
4. Fix `"displays identifier for sent for main sns account"` to expect
the destination to be the subaccount rather than the main account. It
didn't fail because the main account identifier is a substring of the
subaccount identifier (for ICRC-1) and the test used `toContain` instead
of `toBe`.
5. Fix `"displays identifier for sent for sub sns account"` such that
the transaction is actually a sent transaction as the test description
suggests. Also test that it's a sent transaction which the test didn't
test before.

# Tests

Only tests

# Todos

- [ ] Add entry to changelog (if necessary).
Not necessary