Add mapCkbtcTransaction to decode cbor memo (#3735)
# Motivation

In the ckBTC wallets, we want to render burn transactions as withdrawal
transactions.
To accomodate this, https://github.com/dfinity/nns-dapp/pull/3701 adds a
`mapTransaction` parameter to `IcrcTransactionsList` such that we can
use a different transaction mapping function per token.
This PR introduces the alternative mapping function for ckBTC.
But it is not yet used. It will be used in a follow-up PR.

# Changes

1. Install `cbor-web` to be able to decode cbor encoded memos on ckBTC
burn transactions.
2. Add `mapCkbtcTransaction` which behaves like `mapIcrcTransaction` for
most transactions but decodes the withdrawal address from the memo for
burn transactions.

# Tests

1. Add `createBurnTransaction` to conveniently create burn transactions
with specific properties.
2. Create a function `testMapTransactionCommon` which adds all the tests
that are identical between `mapIcrcTransaction` and
`mapCkbtcTransaction` to a `describe` block.
3. Add specific additional tests for burn transactions.

# Todos

- [ ] Add entry to changelog (if necessary).
Will add when it's used.