Fix add-paths lines (#3742)
# Motivation
The snsdemo repo was added accidentally. (Not the whole repo but as a
link.)

This looks as if it was caused by an incorrect pattern.

# Changes
- Remove the spurious directory.
- Fix incorrect patterns.

# Tests
We can trigger workflows but given that the workflows do nothing most of
the time, the only real test will be to wait for changes that cause
non-trivial changes and check whether the bot creates correct outputs.

# Todos

- [x] Add entry to changelog (if necessary).