Render BTC withdrawal address (#3751)
# Motivation

For ckBTC -> BTC withdrawals we want to list the withdrawal address in
the "To:" of the transactions.
The withdrawals address in encoded in the memo of the ckBTC burn
transaction.
We already have a custom `mapCkbtcTransaction` function which can be
used instead of `mapIcrcTransaction` to map the transaction such that if
it is a burn transaction with a memo, the decoded address will be set in
the "To:".

In `TransactionCard.svelte` we render either To/Source for a Transfer or
a custom description for either transaction types. We never render both,
but technically `TransactionCard` currently supports rendering both,
even though it doesn't look right in the UI.

# Changes

1. Pass `mapCkbtcTransaction` instead of `mapIcrcTransaction` to
`IcrcTransactionsList` from `fCkBTCTransactionsList.svelte` to take care
of decoding the withdrawal address.
2. In `TransactionCard.svelte`, only render a description if no
"To:/Source:" identifier is present.

# Tests

1. Test that ckBTC burn transactions without memo render as before.
2. Test that ckBTC burn transactions with correctly encoded memo render
the encoded withdrawal address.
3. Test that a generic burn transaction without to/from is rendered with
the given description.
4. Test that a generic burn transaction with a `to` gets rendered with
the "To:"

Manually at:
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Todos

- [x] Add entry to changelog (if necessary).