Create UiTransactions in IcrcTransactionsList (#3799)
# Motivation

To simplify transactions list rendering and give the higher level
components more control over it, I'm moving the conversion to
`UiTransaction` up the component chain.
The ultimate goal is to create the list of `UiTransaction`s directly
from the list of `TransactionWithId`s without intermediate
`IcrcTransactionData` or `Transaction`. Then it will also be easier to
render multiple transactions as 1 (to hide the approval step) or 1 as
multiple (such as with a to-self transaction).

In this PR we move creation of `UiTransactions` from the
IcrcTransactionCard` to the `IcrcTransactionsList`.

# Changes

1. Add a field `domKey` to `UiTransaction` to keep consistent list
rendering.
2. Pass transaction ID to `toUiTransaction` to use for the `domKey`.
3. Remove all props from `IcrcTransactionCard` and keep only 1 prop:
`transaction: UiTransaction`.
4. In `IcrcTransactionsList` convert the entire list of transactions to
a list of `UiTransaction`s before rendering.

# Tests

1. Remove most tests from `IcrcTransactionCard.spec.ts` because the
component became a very simple pass-through components.
2. Add a test that was missing that no transactions are rendered when
`token` is not available yet.
3. Add `domKey` to unit tests.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary