GIX-2085: AccountsPlusPagePo (#3806)
# Motivation

There was a mismatch with the PO used in one test and the actual
component being tested.

The spec for the `+page` in the accounts route was using the PO for the
`routes/Accounts.svelte`. Which is really a child of that component.

In this PR, I add a new PO `AccountsPlusPagePo`. I added the `PlusPage`
in reference to `+page` used in the route. I hope this will help with
the misconcenption.

# Changes

* New PO AccountsPlusPagePo
* Use the new PO in `routes/app/accounts/page.spec.ts`

# Tests

Only test changes.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.