Chore: Add missing fields to SnsSummarySwap (#3827)
# Motivation

The SnsSummarySwap mirrors the `Swap` type in sns-js.

There were two (unused) fields missing in `SnsSummarySwap`.

The aggregator converter was adding them, but because the values in the
`.toEqual(...)` were not typed, it didn't care that there were extra
fields.

I found that out when I put the `toEqual` inside the `SnsSummaryWrapper`
from #3825

# Changes

* Add two missing fields in `SnsSummarySwap`.

# Tests

* Add two fields in the `SnsSummarySwap` mock.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.