GIX-2062: Add loading balance (#3830)
# Motivation

Give a better user experience in the tokens table pages.

In this PR, I introduce two new types `UserTokenBase` and
`UserTokenLoading`. Both are accepted in the props of TokenTable. But
only the initial one `UserTokenData` can be used to dispatch actions.

I guess the loading could also dispatch events, but we should use
another event name, not `nnsAction`. This way, we know for certain that
when `nnsAction` is triggered, it has the token and fee available. This
is important to open modals at the moment, but could be important for
other actions in the future.

# Changes

* Add two new types `UserTokenBase` and `UserTokenLoading`.
* Render a Spinner if UserTokenLoading in TokensTableRow.
* Support a union of `UserTokenLoading` and `UserTokenData` in
TokensTable and TokensTableRow

# Tests

* Add a test in TokensTable.spec for the new loading value

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.