GIX-2121: Fix snsOnlyProjectStore (#3890)
# Motivation

snsOnlyProjectStore should not return universes of ICRC Tokens. It's
used in specifically SNS only places.

# Changes

* Fix `snsOnlyProjectStore` to check whether the current universe is an
ICRC Token universe.

I know that SNS universes are actually ICRC Tokens. But in our current
code, that is not the case, and until we refactor to use the same, we
consider them differently.

# Tests

* Add a test case for `snsOnlyProjectStore`.

Test was failing without the fix.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary yet.